+++ projects/ui-components/lf-repository-browser/lf-repository-browser.component.spec.ts
@@ -117,7 +117,7 @@ describe('LfRepositoryBrowserComponent', () => {
     'getFolderChildrenAsync',
     'getRootTreeNodeAsync',
     'getParentTreeNodeAsync',
-    'getTreeNodeByIdAsync',
+    'getTreeNodeByIdentifierAsync',
   ]);
   let matDialogMock: MatDialog;
   let ngZoneMock: NgZone;